requirements.txt
groq>=0.4.0
python-dotenv>=1.0.1
rich>=13.0.0
